presenter/pyproject.toml
[tool.poetry]
name = "presenter"
version = "0.1.0"
description = "Data presenter"
authors = ["Maciej Bociański"]
license = "Cognizant"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
